pc-tty: extend capacity of CSI params
Allow CSI parameters to be in 16bit range instead of 8bit value.
Some of the CSI parameters used by e.g. `vi` may overflow 8bit
values while reading `CSI ? 1049 l/h`, 1049 is a parameter
grater than 8bit value.

JIRA: RTOS-378

+++ tty/pc-tty/ttypc_vt.h
@@ -125,7 +125,7 @@ typedef struct {
 	uint8_t dcsst;           /* Device Control String state machine */
 	uint8_t escst;           /* Escape sequence state machine */
 	uint8_t parmi;           /* Escape sequence parameter index */
-	uint8_t parms[MAXPARMS]; /* Escape sequence parameter array */
+	uint16_t parms[MAXPARMS]; /* Escape sequence parameter array */
 	char tabs[MAXTABS];      /* Table of active tab stops */
 	libtty_common_t tty;     /* Terminal character processing (using libtty) */
